Gewichtsangaben vereinheitlichen

+++ src/heroes/inventory/CloseCombatWeaponEditor.java
@@ -47,7 +47,7 @@ public class CloseCombatWeaponEditor {
 	@FXML
 	private Button okButton;
 	@FXML
-	private ReactiveSpinner<Integer> weight;
+	private ReactiveSpinner<Double> weight;
 	@FXML
 	private ReactiveSpinner<Integer> ini;
 	@FXML
@@ -138,7 +138,7 @@ public class CloseCombatWeaponEditor {
 			tpkkThreshold.getValueFactory().setValue(tpkkValues._1);
 			tpkkStep.getValueFactory().setValue(tpkkValues._2);
 		}
-		weight.getValueFactory().setValue((int) weapon.getWeight());
+		weight.getValueFactory().setValue(weapon.getWeight());
 		length.getValueFactory().setValue(weapon.getLength());
 		noBf.setSelected(weapon.getBf() != Integer.MIN_VALUE);
 		if (weapon.getBf() == Integer.MIN_VALUE) {
